requirements.txt
annotated-types==0.7.0
ansi2html==1.8.0
anyio==3.7.1
blinker==1.7.0
build==1.0.3
cachelib==0.9.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.5
dash==2.15.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-extensions==1.0.12
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-leaflet==1.0.15
dash-table==5.0.0
dash_mantine_components==0.14.4
dataclass-wizard==0.22.2
deprecation==2.1.0
dill==0.3.7
diskcache==5.6.1
ecdsa==0.19.0
EditorConfig==0.12.3
fastapi==0.109.2
Flask==3.0.2
Flask-Caching==2.1.0
geobuf==1.1.1
h11==0.14.0
idna==3.4
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.4
jsbeautifier==1.14.11
MarkupSafe==2.1.3
more-itertools==9.1.0
multiprocess==0.70.15
nest-asyncio==1.5.6
numpy==1.26.4
packaging==23.1
pandas==2.0.3
parameterized==0.9.0
pip-check==2.8.1
pip-tools==7.4.0
pipdeptree==2.13.0
plotly==5.15.0
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==3.3.0
pytz==2023.3
PyYAML>=6.0
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette
tenacity==8.2.2
terminaltables==3.1.10
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
uuid==1.30
uvicorn==0.23.0
Werkzeug==3.0.4
zipp==3.20.1
